Bring back old tests to test deprecated classes
